use dliess/cppzmq

+++ tests/TemperatureQt/Server.cpp
@@ -85,7 +85,7 @@ int main()
 
     utils::FdSet fdSet;
     fdSet.AddFd(timerFd, [&](int fd){
-        int timersElapsed = 0;
+        int64_t timersElapsed = 0;
         (void) read(fd, &timersElapsed, 8);
         float diff = commandedTemperature - actualTemperatureTmp;
         if(std::abs(diff) > 0.005)
